Moved visualization scripts and sample scenes to the Samples~ folder
