--- Runtime/Scripts/Visualizations.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: fe7a416eae864abaaabf455e6b81bbe3
-timeCreated: 1658764737
