+++ Samples~/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 41747461d3d6923408034ba16d51c83d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
